substantial changes to ComputeExpectedValueJoint
ComputeExpectedValueJoint is now separated into
Observer.UpdateExperience (which returns a PC object with samples from
an agent who may be potentially updating beliefs of their costs and
rewards) and AuxiliaryFunctions.ProbabilityOfChange() which returns the
probability that any given cost or reward variable was updated in the
agent’s mind given the original PC and the updated PC objects. The
functions are split because you can then store the second PC and query
ProbabilityOfChange() quickly about different variables.
Observer.ComputeProbabiligyOfChange() runs both functions sequentially
and returns the probably (but the PC from the UpdateExperience() method
gets destroyed).