Observer.SimulateAgents() can now be run without planning
The replan boolean flag now tells the function whether to compute the policy or not. This lets the code run much faster when you want to draw multiple samples from the same agent. The disadvantage is that when you initialize things you'll have an empty policy (is this true?) so you'll get nonsense. To be safe you can explicity build the policy through O.Plr.Prepare() and then running SimulateAgents with no replaning will always work.